Split out ConsoleUtils
